chore: Make Presentation lib compile.
