+++ src/app/ApplicationTemplate.Client/Framework/Serialization/ObjectSerializerToContentSerializerAdapter.cs
