+++ src/app/ApplicationTemplate.Client/Framework/Serialization/ObjectSerializerToSettingsSerializerAdapter.cs
