+++ src/app/ApplicationTemplate.Client/Framework/Serialization/SystemTextJsonToContentSerializerAdapter.cs
