+++ src/app/ApplicationTemplate.Client/Framework/Serialization/SystemTextJsonToObjectSerializerAdapter.cs
