+++ src/app/ApplicationTemplate.Presentation/Framework/Configuration/IConfigurationExtensions.cs
