+++ src/app/ApplicationTemplate.Presentation/Framework/DataLoader/AnalyticsDataLoaderStrategy.cs
