+++ src/app/ApplicationTemplate.Presentation/Framework/DependencyInjection/RegistrationServiceCollectionExtensions.cs
