+++ src/app/ApplicationTemplate.Presentation/Framework/Reactive/IObservable.Extensions.cs
