+++ src/app/ApplicationTemplate.Presentation/Framework/Startup/CoreStartupBase.cs
