+++ src/app/ApplicationTemplate.Presentation/Framework/Startup/StartupState.cs
